pyproject.toml
[tool.poetry]
name = "sae_feature_scouter"
version = "0.1.0"
description = "A scalable, production-ready SAE feature scouter toolkit for mechanistic interpretability research."
authors = ["Adyn Blaed <adyn@blaed.net>"]
readme = "README.md"
license = "MIT"
homepage = "https://github.com/adynblaed"
repository = "https://github.com/adynblaed"
documentation = "https://github.com/adynblaed"
keywords = ["mechanistic-interpretability", "sae", "feature-extraction", "visualization"]
classifiers = [
    "Development Status :: v0.1.0",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
uvicorn = "^0.34.0"
fastapi = "^0.115.8" 
huggingface-hub = "^0.28.1"
sae_lens = "^5.4.2"
sae-vis = "^0.3.6"
datasets = "2.21.0"
pyzmq = "^26.0.0"
zstandard = "0.22.0"
frozendict = "^2.4.6"
tabulate = "^0.9.0"
torch = "^2.6.0"
torchvision = "^0.21.0"
torchaudio = "^2.6.0"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.0"
pytest = "^8.3.4"
ruff = "^0.9.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88
